ISIS-2717: improves fixture scripts docs; moves classes around
