+++ antora/components/refguide-index/modules/testing/pages/index/fixtures/applib/clock/clock/Clock.adoc
@@ -1,0 +1,80 @@
+= Clock
+:Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+
+Provides a mechanism to get (and possible to set) the current time.
+
+The clock is used primarily by the temporal value classes, and is accessed by the framework as a singleton. The actual implementation used can be configured at startup, but once specified the clock instance cannot be changed.
+
+Unless another xref:refguide:testing:index/fixtures/applib/clock/clock/Clock.adoc[Clock] implementation has been installed, the first call to _#getInstance()_ will instantiate an implementation that just uses the system's own clock. Alternate implementations can be created via suitable subclasses, but this must be done * _before_ * the first call to _#getInstance()_ .
+
+== API
+
+[source,java]
+.Clock.java
+----
+class Clock {
+  Clock getInstance()     // <.>
+  boolean isInitialized()     // <.>
+  long getEpochMillis()     // <.>
+  LocalDate getTimeAsLocalDate()
+  LocalDateTime getTimeAsLocalDateTime()
+  OffsetDateTime getTimeAsOffsetDateTime()     // <.>
+  Timestamp getTimeAsJavaSqlTimestamp()
+  DateTime getTimeAsJodaDateTime()     // <.>
+}
+----
+
+<.> xref:#getInstance__[getInstance()]
++
+--
+Returns the (singleton) instance of xref:refguide:testing:index/fixtures/applib/clock/clock/Clock.adoc[Clock] .
+--
+<.> xref:#isInitialized__[isInitialized()]
++
+--
+Whether has been initialized or not.
+--
+<.> xref:#getEpochMillis__[getEpochMillis()]
++
+--
+The time as the number of milliseconds since the epoch start. (UTC)
+--
+<.> xref:#getTimeAsOffsetDateTime__[getTimeAsOffsetDateTime()]
++
+--
+Returns the time as _OffsetDateTime_ , using the _ZoneId#systemDefault() system default_ timezone.
+--
+<.> xref:#getTimeAsJodaDateTime__[getTimeAsJodaDateTime()]
++
+--
+Returns the time as a Joda _DateTime_ , using the _ZoneId#systemDefault() system default_ timezone.
+--
+
+== Members
+
+[#getInstance__]
+=== getInstance()
+
+Returns the (singleton) instance of xref:refguide:testing:index/fixtures/applib/clock/clock/Clock.adoc[Clock] .
+
+Unless it has been otherwise created, will lazily instantiate an implementation that just delegate to the computer's own system clock (as per _System#currentTimeMillis()_ .
+
+[#isInitialized__]
+=== isInitialized()
+
+Whether has been initialized or not.
+
+[#getEpochMillis__]
+=== getEpochMillis()
+
+The time as the number of milliseconds since the epoch start. (UTC)
+
+[#getTimeAsOffsetDateTime__]
+=== getTimeAsOffsetDateTime()
+
+Returns the time as _OffsetDateTime_ , using the _ZoneId#systemDefault() system default_ timezone.
+
+[#getTimeAsJodaDateTime__]
+=== getTimeAsJodaDateTime()
+
+Returns the time as a Joda _DateTime_ , using the _ZoneId#systemDefault() system default_ timezone.
