+++ antora/components/refguide-index/modules/testing/pages/index/fixtures/applib/clock/clock/FixtureClock.adoc
@@ -1,0 +1,117 @@
+= FixtureClock
+:Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+
+This clock, for use by fixtures, can be set to specific time.
+
+If not set it will provide the time provided by the system clock.
+
+Note that - by design - it does not provide any mechanism to advance the time (eg automatic ticking of the clock). That is, the time returned is always explicitly under the control of the programmer (it can be moved forward or back as required).
+
+== API
+
+[source,java]
+.FixtureClock.java
+----
+class FixtureClock {
+  FixtureClock initialize()     // <.>
+  boolean remove()     // <.>
+  void clear()     // <.>
+  void setTime(final int hour, final int min)     // <.>
+  void setDate(final int year, final int month, final int day)     // <.>
+  void addTime(final int hours, final int minutes)     // <.>
+  void addDate(final int years, final int months, final int days)     // <.>
+  void reset()     // <.>
+  String toString()
+}
+----
+
+<.> xref:#initialize__[initialize()]
++
+--
+Configures the system to use a FixtureClock rather than the in-built system clock. Can be called multiple times.
+--
+<.> xref:#remove__[remove()]
++
+--
+Makes _Clock#remove()_ visible.
+--
+<.> xref:#clear__[clear()]
++
+--
+Sets the clock to epoch, that is midnight, 1 Jan 1970 UTC.
+--
+<.> xref:#setTime__int_int[setTime(int, int)]
++
+--
+Sets the hours and minutes as specified, and sets the seconds and milliseconds to zero, but the date portion is left unchanged.
+--
+<.> xref:#setDate__int_int_int[setDate(int, int, int)]
++
+--
+Sets the date, but the time portion is left unchanged.
+--
+<.> xref:#addTime__int_int[addTime(int, int)]
++
+--
+Adjusts the time by the specified number of hours and minutes.
+--
+<.> xref:#addDate__int_int_int[addDate(int, int, int)]
++
+--
+Adjusts the time by the specified number of years, months or days.
+--
+<.> xref:#reset__[reset()]
++
+--
+Go back to just returning the system's time.
+--
+
+== Members
+
+[#initialize__]
+=== initialize()
+
+Configures the system to use a FixtureClock rather than the in-built system clock. Can be called multiple times.
+
+Must call before any other call to _Clock#getInstance()_ .
+
+[#remove__]
+=== remove()
+
+Makes _Clock#remove()_ visible.
+
+[#clear__]
+=== clear()
+
+Sets the clock to epoch, that is midnight, 1 Jan 1970 UTC.
+
+This is typically called before either _#setDate(int, int, int)_ (so that time is set to midnight) and/or _#setTime(int, int)_ (so that date is set to a well known value).
+
+[#setTime__int_int]
+=== setTime(int, int)
+
+Sets the hours and minutes as specified, and sets the seconds and milliseconds to zero, but the date portion is left unchanged.
+
+[#setDate__int_int_int]
+=== setDate(int, int, int)
+
+Sets the date, but the time portion is left unchanged.
+
+[#addTime__int_int]
+=== addTime(int, int)
+
+Adjusts the time by the specified number of hours and minutes.
+
+Typically called after _#setTime(int, int)_ , to move the clock forward or perhaps back.
+
+[#addDate__int_int_int]
+=== addDate(int, int, int)
+
+Adjusts the time by the specified number of years, months or days.
+
+Typically called after _#setDate(int, int, int)_ , to move the clock forward or perhaps back.
+
+[#reset__]
+=== reset()
+
+Go back to just returning the system's time.
