+++ antora/components/refguide-index/modules/testing/pages/index/fixtures/applib/fixturescripts/FixtureScriptWithExecutionStrategy.adoc
@@ -1,17 +1,17 @@
-= BuilderScriptAbstract
+= FixtureScriptWithExecutionStrategy _(interface)_
 :Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+
+Interface for xref:refguide:testing:index/fixtures/applib/fixturescripts/FixtureScript.adoc[FixtureScript] s to optionally implement, used to override the _FixtureScriptsSpecification#getMultipleExecutionStrategy() globally-defined_ _org.apache.isis.testing.fixtures.applib.fixturescripts.FixtureScripts.MultipleExecutionStrategy_ strategy.
+
+This therefore allows individual fixture scripts to indicate that they have their own execution strategy.
 
 == API
 
 [source,java]
-.BuilderScriptAbstract.java
+.FixtureScriptWithExecutionStrategy.java
 ----
-class BuilderScriptAbstract<T> {
-  BuilderScriptAbstract<T> build(final FixtureScript parentFixtureScript, final ExecutionContext executionContext)
-  T objectFor(final PersonaWithBuilderScript<BuilderScriptAbstract<T>> persona, final ExecutionContext executionContext)
-  T findUsing(final PersonaWithFinder<T> persona)
-  BuilderScriptAbstract<T> setPrereq(WithPrereqs.Block<T> prereq)
-  void execPrereqs(final ExecutionContext executionContext)
+interface FixtureScriptWithExecutionStrategy {
+  FixtureScripts.MultipleExecutionStrategy getMultipleExecutionStrategy()
 }
 ----
