+++ antora/components/refguide-index/modules/testing/pages/index/fixtures/applib/personas/BuilderScriptAbstract.adoc
@@ -1,0 +1,21 @@
+= BuilderScriptAbstract
+:Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+
+A specialization of xref:refguide:testing:index/fixtures/applib/fixturescripts/FixtureScript.adoc[FixtureScript] that is intended to be used to setup the state of a xref:refguide:testing:index/fixtures/applib/personas/PersonaWithBuilderScript.adoc[PersonaWithBuilderScript] .
+
+The xref:refguide:testing:index/fixtures/applib/personas/PersonaWithBuilderScript.adoc[PersonaWithBuilderScript] represents the "what", in other words the raw data, while the xref:refguide:testing:index/fixtures/applib/personas/BuilderScriptAbstract.adoc[BuilderScriptAbstract] represents the "how-to".
+
+== API
+
+[source,java]
+.BuilderScriptAbstract.java
+----
+class BuilderScriptAbstract<T> {
+  BuilderScriptAbstract<T> build(final FixtureScript parentFixtureScript, final ExecutionContext executionContext)
+  T objectFor(final PersonaWithBuilderScript<BuilderScriptAbstract<T>> persona, final ExecutionContext executionContext)
+  T findUsing(final PersonaWithFinder<T> persona)
+  BuilderScriptAbstract<T> addPrereq(WithPrereqs.Block<T> prereq)
+  void execPrereqs(final ExecutionContext executionContext)
+}
+----
+
