+++ antora/components/refguide-index/modules/testing/pages/index/fixtures/applib/personas/BuilderScriptWithResult.adoc
@@ -1,0 +1,28 @@
+= BuilderScriptWithResult
+:Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+
+A specialization of xref:refguide:testing:index/fixtures/applib/personas/BuilderScriptAbstract.adoc[BuilderScriptAbstract] which expects there to be a top-level object, and so - by defining its own _#buildResult(ExecutionContext) hook method_ - removes a little of the boilerplate..
+
+== API
+
+[source,java]
+.BuilderScriptWithResult.java
+----
+class BuilderScriptWithResult<T> {
+  public T object;
+  T getObject()     // <.>
+}
+----
+
+<.> xref:#getObject__[getObject()]
++
+--
+Simply returns the object returned by _#buildResult(ExecutionContext)_ .
+--
+
+== Members
+
+[#getObject__]
+=== getObject()
+
+Simply returns the object returned by _#buildResult(ExecutionContext)_ .
