+++ antora/components/refguide-index/modules/testing/pages/index/fixtures/applib/personas/PersonaWithBuilderScript.adoc
@@ -1,0 +1,29 @@
+= PersonaWithBuilderScript _(interface)_
+:Notice: Licensed to the Apache Software Foundation (ASF) under one or more contributor license agreements. See the NOTICE file distributed with this work for additional information regarding copyright ownership. The ASF licenses this file to you under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at. http://www.apache.org/licenses/LICENSE-2.0 . Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR  CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
+
+Intended for persona enums to implement, to provide an instance of a xref:refguide:testing:index/fixtures/applib/personas/BuilderScriptAbstract.adoc[BuilderScriptAbstract] in order to instantiate an instance of the persona (normally in the form of a domain entity or set of related domain entities).
+
+( xref:refguide:testing:index/fixtures/applib/personas/BuilderScriptAbstract.adoc[BuilderScriptAbstract] is a specialization of _org.apache.isis.testing.fixtures.applib.fixturescripts.FixtureScript_ ).
+
+== API
+
+[source,java]
+.PersonaWithBuilderScript.java
+----
+interface PersonaWithBuilderScript<T> {
+  T builder()     // <.>
+}
+----
+
+<.> xref:#builder__[builder()]
++
+--
+Returns a xref:refguide:testing:index/fixtures/applib/personas/BuilderScriptAbstract.adoc[BuilderScriptAbstract] to use to instantiate this persona.
+--
+
+== Members
+
+[#builder__]
+=== builder()
+
+Returns a xref:refguide:testing:index/fixtures/applib/personas/BuilderScriptAbstract.adoc[BuilderScriptAbstract] to use to instantiate this persona.
