+++ testing/fixtures/applib/src/test/java/org/apache/isis/testing/fixtures/applib/legacy/FixtureClockInstantiationTest.java
@@ -27,8 +27,8 @@ import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
-import org.apache.isis.testing.fixtures.applib.clock.Clock;
-import org.apache.isis.testing.fixtures.applib.clock.FixtureClock;
+import org.apache.isis.testing.fixtures.applib.clock.clock.Clock;
+import org.apache.isis.testing.fixtures.applib.clock.clock.FixtureClock;
 
 public class FixtureClockInstantiationTest {
